requirements.txt
# Create automatically: $ pipreqs .

google_api_python_client==2.62.0
pip==22.2.2
protobuf==4.21.6
python_dateutil==2.8.2
